Initial commit for BOSH release
This BOSH release includes a single package with a single binary file.
That file is executable and will start a process that can be accessed
via HTTP GET at the port provided by the first argument.

The port is provided via the default $PORT environment variable in the
go-http-helloworld_ctl file.